Project1-Financial_Autoagentic_System/pyproject.toml
[project]
name = "autoagent"
version = "0.1.0"
description = "Self-improving financial research agents using the autoresearch pattern"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "httpx>=0.27.0",
    "edgartools>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
phase2 = [
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.4.0",
]